// src/app/pages/external-reviewers/types/acceptance-criteria/external-reviewer-acceptance-criterion-attachment.ts
export class ExternalReviewerAcceptanceCriterionAttachment {
    id?: number;
    attachmentNameAr?: string;
    attachmentNameEn?: string;
    attachmentFileBucketName?: string;
    attachmentFileObjectName?: string;
    notes?: string;
    module?: string;
    createdBy?: number;
    createdOn?: string;
    updatedBy?: number;
    updatedOn?: string;
    user?: any;
}